+++ catalogue-api/src/main/java/io/quartic/catalogue/api/AbstractDatasetConfig.java
@@ -1,0 +1,11 @@
+package io.quartic.catalogue.api;
+
+import io.quartic.weyl.common.SweetStyle;
+import org.immutables.value.Value;
+
+@SweetStyle
+@Value.Immutable
+public interface AbstractDatasetConfig {
+    DatasetMetadata metadata();
+    DatasetLocator locator();
+}
