+++ catalogue-api/src/main/java/io/quartic/catalogue/api/AbstractDatasetId.java
@@ -1,0 +1,10 @@
+package io.quartic.catalogue.api;
+
+import io.quartic.weyl.common.SweetStyle;
+import io.quartic.weyl.common.uid.Uid;
+import org.immutables.value.Value;
+
+@SweetStyle
+@Value.Immutable
+public interface AbstractDatasetId extends Uid {
+}
