+++ catalogue-api/src/main/java/io/quartic/catalogue/api/AbstractIcon.java
@@ -1,0 +1,12 @@
+package io.quartic.catalogue.api;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+import io.quartic.weyl.common.SweetStyle;
+import org.immutables.value.Value;
+
+@SweetStyle
+@Value.Immutable
+public interface AbstractIcon {
+    @JsonValue
+    String icon();
+}
