+++ catalogue-api/src/main/java/io/quartic/catalogue/api/AbstractPostgresDatasetLocator.java
@@ -1,0 +1,13 @@
+package io.quartic.catalogue.api;
+
+import io.quartic.weyl.common.SweetStyle;
+import org.immutables.value.Value;
+
+@SweetStyle
+@Value.Immutable
+public interface AbstractPostgresDatasetLocator extends DatasetLocator {
+    String user();
+    String password();
+    String url();
+    String query();
+}
