+++ catalogue-api/src/main/java/io/quartic/catalogue/api/AbstractWebsocketDatasetLocator.java
@@ -1,0 +1,10 @@
+package io.quartic.catalogue.api;
+
+import io.quartic.weyl.common.SweetStyle;
+import org.immutables.value.Value;
+
+@SweetStyle
+@Value.Immutable
+public interface AbstractWebsocketDatasetLocator extends DatasetLocator {
+    String url();
+}
