+++ catalogue-application/src/main/java/io/quartic/catalogue/CatalogueConfiguration.java
@@ -1,0 +1,6 @@
+package io.quartic.catalogue;
+
+import io.dropwizard.Configuration;
+
+public class CatalogueConfiguration extends Configuration {
+}
