+++ catalogue-application/src/test/resources/catalogue.yml
