--- jester-api/src/main/java/io/quartic/jester/api/AbstractDatasetConfig.java
@@ -1,11 +1,0 @@
-package io.quartic.jester.api;
-
-import io.quartic.weyl.common.SweetStyle;
-import org.immutables.value.Value;
-
-@SweetStyle
-@Value.Immutable
-public interface AbstractDatasetConfig {
-    DatasetMetadata metadata();
-    DatasetLocator locator();
-}
