--- jester-api/src/main/java/io/quartic/jester/api/AbstractDatasetId.java
@@ -1,10 +1,0 @@
-package io.quartic.jester.api;
-
-import io.quartic.weyl.common.SweetStyle;
-import io.quartic.weyl.common.uid.Uid;
-import org.immutables.value.Value;
-
-@SweetStyle
-@Value.Immutable
-public interface AbstractDatasetId extends Uid {
-}
