--- jester-api/src/main/java/io/quartic/jester/api/AbstractDatasetMetadata.java
@@ -1,15 +1,0 @@
-package io.quartic.jester.api;
-
-import io.quartic.weyl.common.SweetStyle;
-import org.immutables.value.Value;
-
-import java.util.Optional;
-
-@SweetStyle
-@Value.Immutable
-public interface AbstractDatasetMetadata {
-    String name();
-    String description();
-    String attribution();
-    Optional<Icon> icon();
-}
