--- jester-api/src/main/java/io/quartic/jester/api/AbstractIcon.java
@@ -1,12 +1,0 @@
-package io.quartic.jester.api;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-import io.quartic.weyl.common.SweetStyle;
-import org.immutables.value.Value;
-
-@SweetStyle
-@Value.Immutable
-public interface AbstractIcon {
-    @JsonValue
-    String icon();
-}
