--- jester-api/src/main/java/io/quartic/jester/api/AbstractPostgresDatasetLocator.java
@@ -1,13 +1,0 @@
-package io.quartic.jester.api;
-
-import io.quartic.weyl.common.SweetStyle;
-import org.immutables.value.Value;
-
-@SweetStyle
-@Value.Immutable
-public interface AbstractPostgresDatasetLocator extends DatasetLocator {
-    String user();
-    String password();
-    String url();
-    String query();
-}
