--- jester-application/build.gradle
@@ -1,43 +1,0 @@
-plugins {
-    id "java"
-    id "idea"
-    id "application"
-    id "org.inferred.processors" version "1.2.3"
-    id "com.palantir.docker" version "0.9.1"
-    id "com.palantir.configuration-resolver" version "0.1.0"
-}
-
-dependencies {
-    processor libraries.immutables
-
-    compile project(":common-application")
-    compile project(":jester-api")
-
-    testCompile libraries.testCommon
-    testCompile "io.dropwizard:dropwizard-testing:${versions.dropwizard}"
-}
-
-jar {
-    manifest {
-        attributes("Implementation-Version": version)
-    }
-}
-
-ext.memorySetting = System.env.JESTER_MEMORY ?: "64m"
-
-mainClassName = "io.quartic.jester.JesterApplication"
-applicationDefaultJvmArgs = ["-Xms${memorySetting}", "-Xmx${memorySetting}", "-XX:+PrintGCDetails", "-XX:+PrintGCDateStamps", "-Xloggc:mygclogfilename.gc"]
-
-
-run {
-    args = ["server", "jester.yml"]
-}
-
-docker {
-    name      "${System.env.QUARTIC_DOCKER_REPOSITORY}/jester:${version}"
-    dependsOn distTar
-    files     "build.gradle"    // Seems to be no way to say "no files", so using this as dummy
-}
-
-
-
