--- jester-application/src/main/java/io/quartic/jester/JesterConfiguration.java
@@ -1,6 +1,0 @@
-package io.quartic.jester;
-
-import io.dropwizard.Configuration;
-
-public class JesterConfiguration extends Configuration {
-}
