+++ settings.gradle
@@ -1,7 +1,7 @@
 include "common-core"
 include "common-application"
-include "jester-api"
-include "jester-application"
+include "catalogue-api"
+include "catalogue-application"
 include "weyl-core"
 include "weyl-application"
 include "weyl-frontend"
