+++ weyl-application/src/main/java/io/quartic/weyl/WeylConfiguration.java
@@ -3,13 +3,13 @@ package io.quartic.weyl;
 import io.dropwizard.Configuration;
 
 public class WeylConfiguration extends Configuration {
-    private String jesterUrl;
+    private String catalogueUrl;
 
-    public String getJesterUrl() {
-        return jesterUrl;
+    public String getCatalogueUrl() {
+        return catalogueUrl;
     }
 
-    public void setJesterUrl(String jesterUrl) {
-        this.jesterUrl = jesterUrl;
+    public void setCatalogueUrl(String catalogueUrl) {
+        this.catalogueUrl = catalogueUrl;
     }
 }
