+++ weyl-core/build.gradle
@@ -8,7 +8,7 @@ dependencies {
     processor libraries.immutables
 
     compile project(":common-core")
-    compile project(":jester-api")  // TODO: this is probably wrong - should move the dataset types to common
+    compile project(":catalogue-api")  // TODO: this is probably wrong - should move the dataset types to common
 
     compile "org.postgresql:postgresql:9.4.1209"
     compile "org.apache.commons:commons-io:1.3.2"
